pyproject.toml
[tool.poetry]
name = "papers-fetcher"
version = "0.1.0"
description = "A tool to fetch research papers from PubMed with pharmaceutical/biotech company affiliations"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
requests = "^2.31.0"
typing-extensions = "^4.7.1"
pandas = "^2.0.3"
biopython = "^1.81"
typer = "^0.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
mypy = "^1.3.0"
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
get-papers-list = "cli.main:main"